#76 Рефакторинг поведения игрока
Отдельные поведения вынесены в соответсвующие классы, обработка инвентаря немного изменена (теперь возвращаем реакцию и выполняем ее в поведении игрока (заготовка для ситуации где надо сначала цель выбрать, а потом применять реакцию))

Так же убран дебаг-мод верстки, и изменен заголовок окна игры

Требуется рефакторинг поведений игрока
