+++ game/src/main/java/com/vdn/lampbearer/config/GameConfig.java
@@ -2,7 +2,6 @@ package com.vdn.lampbearer.config;
 
 import org.hexworks.zircon.api.CP437TilesetResources;
 import org.hexworks.zircon.api.application.AppConfig;
-import org.hexworks.zircon.api.application.DebugConfig;
 import org.hexworks.zircon.api.component.ColorTheme;
 
 public class GameConfig {
@@ -11,18 +10,19 @@ public class GameConfig {
     public static final int WINDOW_WIDTH = 100;
     public static final int WINDOW_HEIGHT = 50;
     public static final int SIDEBAR_WIDTH = 30;
-    public static final int SIDEBAR_HEIGHT = 50;
+    public static final int SIDEBAR_HEIGHT = 49;
     public static final int LOG_AREA_HEIGHT = 10;
 
     public static AppConfig getAppConfig() {
         return AppConfig.newBuilder()
                 .withDefaultTileset(CP437TilesetResources.anikki16x16())
                 .withSize(WINDOW_WIDTH, WINDOW_HEIGHT)
-                .withFpsLimit(60)
-                .withDebugMode(true)
-                .withDebugConfig(DebugConfig.newBuilder()
-                        .withRelaxBoundsCheck(true)
-                        .build())
+                .withTitle("TMIQ III the Lampbearer")
+                .withFpsLimit(30)
+//                .withDebugMode(true)
+//                .withDebugConfig(DebugConfig.newBuilder()
+//                        .withRelaxBoundsCheck(true)
+//                        .build())
                 .build();
     }
 }
