+++ game/src/main/java/com/vdn/lampbearer/views/inventory/ItemActionModalView.java
@@ -1,6 +1,7 @@
 package com.vdn.lampbearer.views.inventory;
 
 import com.vdn.lampbearer.config.GameConfig;
+import com.vdn.lampbearer.dto.ItemUseReactionContextDto;
 import com.vdn.lampbearer.entites.item.AbstractItem;
 import com.vdn.lampbearer.game.GameContext;
 import com.vdn.lampbearer.views.fragments.ItemUseFragment;
@@ -9,10 +10,10 @@ import org.hexworks.zircon.api.Components;
 import org.hexworks.zircon.api.builder.component.ModalBuilder;
 import org.hexworks.zircon.api.component.Button;
 import org.hexworks.zircon.api.component.ComponentAlignment;
+import org.hexworks.zircon.api.fragment.menu.MenuItemSelected;
 import org.hexworks.zircon.api.fragment.menu.SelectionCancelled;
 import org.hexworks.zircon.api.graphics.BoxType;
-import org.hexworks.zircon.internal.component.modal.EmptyModalResult;
-import org.jetbrains.annotations.NotNull;
+import org.jetbrains.annotations.Nullable;
 
 import java.util.concurrent.CountDownLatch;
 import java.util.concurrent.atomic.AtomicReference;
@@ -21,8 +22,8 @@ import static org.hexworks.zircon.api.ComponentDecorations.box;
 import static org.hexworks.zircon.api.ComponentDecorations.noDecoration;
 
 public class ItemActionModalView {
-    @NotNull
-    public static Boolean showItemActionModal(GameContext context, AbstractItem item) {
+    @Nullable
+    public static ItemUseReactionContextDto showItemActionModal(GameContext context, AbstractItem item) {
         var panel = Components.panel()
                 .withPreferredSize(27, item
                         .getActions().size() + 3)
@@ -37,10 +38,7 @@ public class ItemActionModalView {
                 .withColorTheme(GameConfig.THEME)
                 .build();
 
-        ItemUseFragment itemUseFragment = new ItemUseFragment(
-                item,
-                context,
-                modal);
+        ItemUseFragment itemUseFragment = new ItemUseFragment(item, context, modal);
 
         panel.addFragment(itemUseFragment);
 
@@ -59,12 +57,11 @@ public class ItemActionModalView {
 
         CountDownLatch countDownLatch = new CountDownLatch(1);
 
-        AtomicReference<Boolean> isSelectedItemAction = new AtomicReference<>(false);
+        AtomicReference<ItemUseReactionContextDto> selectedItemActionContext =
+                new AtomicReference<>(null);
         modal.onClosed(modalResult -> {
-            if (modalResult instanceof EmptyModalResult) {
-                isSelectedItemAction.set(true);
-            } else {
-                isSelectedItemAction.set(false);
+            if (modalResult instanceof MenuItemSelected) {
+                selectedItemActionContext.set(((MenuItemSelected<ItemUseReactionContextDto>) modalResult).getItem());
             }
             countDownLatch.countDown();
             return Unit.INSTANCE;
@@ -78,6 +75,6 @@ public class ItemActionModalView {
             throw new RuntimeException(e);
         }
 
-        return isSelectedItemAction.get();
+        return selectedItemActionContext.get();
     }
 }
